requirements.txt
astro-sdk-python==1.8.1
astronomer-providers==1.19.4
apache-airflow-providers-airbyte==4.0.0
dbt-core==1.8.7
astronomer-cosmos==1.6.0
dbt-postgres==1.8.2
apache-airflow-providers-postgres==5.13.1
apache-airflow-providers-amazon==9.0.0